requirements.txt
psycopg2-binary==2.9.9
fastapi==0.104.1
uvicorn[standard]==0.24.0
sqlalchemy==2.0.23
gunicorn==21.2.0
python-multipart==0.0.6
python-jose[cryptography]==3.3.0
firebase-admin==6.2.0
python-dotenv==1.0.0
pydantic==2.5.0
cryptography==41.0.7
pytest==9.0.2
pytest-asyncio==0.23.2
httpx==0.24.1
